Windows: Implement a WOW64/ARM64EC offline compiler backend program
This tool will be invoked by the offline compiler frontend to compile
caches for a specific executable given a codemap. It assumes the input
codemap is already validated and split, and will validate if any current
caches are complete for the input codemap and (current FEX config), if not
modules will be recompiled as necessary and any stale cache files being
deleted.

Currently config validation is unimplemented, and if to recompile is
determined purely from codemap timestamps indicating an out of date
cache.